code cleanup, absolute paths removed, and code restructured
